% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcMRMR}
\alias{calcMRMR}
\title{Minimum Redundancy, maximum relevance}
\usage{
calcMRMR(time, status, x, nselect = ncol(x))
}
\arguments{
\item{time}{[\code{numeric}]\cr
Vector of survival times.}

\item{status}{[\code{logical}]\cr
Vector of survival event indicators.}

\item{x}{[\code{matrix[double]}]\cr
Matrix of \code{[observations]x[features]}.}

\item{nselect}{[\code{integer(1)}]\cr
Number of features to score. Default is all features.}
}
\value{
[\code{named numeric}]: Vector of scores, named with column
  names of \code{x}, in order of selection.
}
\description{
Minimum Redundancy, maximum relevance
}

