% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fmrmrClassif}
\alias{fmrmr}
\alias{fmrmrClassif}
\alias{fmrmrRegr}
\alias{fmrmrSurv}
\title{Calculate mRMR feature scores.}
\usage{
fmrmrClassif(classes, x, relevance = "chi.squared", redundance = "pearson",
  combine = "difference", alpha = 1)

fmrmrRegr(y, x, relevance = "rank.correlation", redundance = "pearson",
  combine = "difference", alpha = 1)

fmrmrSurv(time, status, x, relevance = "cindex", redundance = "pearson",
  combine = "difference", alpha = 1)
}
\arguments{
\item{classes}{[\code{factor}]\cr
Classs labels.}

\item{time}{[\code{numeric}]\cr
Vector of length \code{n} providing (possibly right censored) survival times.}

\item{status}{[\code{logical}]\cr
Vector of length \code{n} providing censoring informations: \code{FALSE} for
censored observations, \code{TRUE} for uncensored.}

\item{x}{[\code{matrix(n*p)}]\cr
An matrix of \code{n} observations and \code{p} features/scores.}

\item{relevance}{[\code{character(1)}]\cr
Measure for relevance. Possible values are \dQuote{cindex}.
For \dQuote{cindex}, the relevance is determined by twice the absolute distance
to \code{0.5} (random predictor), resulting in a relevance between 0 and 1.
Default is \dQuote{cindex}.}

\item{redundance}{[\code{character(1)}]\cr
Measure for redundance Possible values are \dQuote{pearson} and \dQuote{mi}.
Default is \dQuote{pearson}.}

\item{combine}{[\code{character(1)}]\cr
How to combine redundance and relevance. Possible values are \dQuote{difference}
and \dQuote{quotient}.
Default is \dQuote{difference}.}

\item{alpha}{[\code{numeric(1)}]\cr
Weighting parameter for both combination methods.
Default is 1.}

\item{y}{[\code{numeric}]\cr
Continuous target values.}
}
\value{
[\code{data.frame}]. Contains the following columns (rows are named with feature names):
  \item{\code{score}}{Combined score}
  \item{\code{rel}}{Relevance score}
  \item{\code{red}}{Redundancy score}
}
\description{
Calculate mRMR feature scores.
}

